cleaning up text for submission.  W2V commented out
