+++ create_model-messing-arround.py
@@ -1,18 +1,24 @@
 import pandas as pd
+import sklearn
 from sklearn.feature_extraction.text import TfidfTransformer ,CountVectorizer, TfidfVectorizer
 from scipy.spatial.distance import cosine
 import numpy as np
 from sklearn.linear_model import LogisticRegression
 from sklearn.multiclass import OneVsRestClassifier
-from sklearn.naive_bayes import MultinomialNB
+from sklearn.naive_bayes import MultinomialNB, GaussianNB
 from sklearn.model_selection import cross_val_score
-from sklearn.ensemble import RandomForestClassifier
+from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier, GradientBoostingClassifier, VotingClassifier
 from sklearn.model_selection import train_test_split
 from sklearn.metrics import classification_report
 from sklearn.externals import joblib
 from nltk import PorterStemmer
 from nltk.corpus import stopwords
-analyzer = CountVectorizer().build_analyzer()
+# analyzer = CountVectorizer().build_analyzer()
+from sklearn.neighbors import KNeighborsClassifier
+from sklearn.pipeline import make_pipeline
+from sklearn.preprocessing import Normalizer, MinMaxScaler
+
+analyzer = TfidfVectorizer().build_analyzer()
 import scipy
 from sklearn.svm import SVC
 import gensim, logging
@@ -20,6 +26,10 @@ from gensim.models.doc2vec import LabeledSentence
 from textblob import TextBlob
 from textblob.sentiments import NaiveBayesAnalyzer
 is_debug = False
+import spacy # Requirements: yum install python-devel python3-dev  libevent-dev; pip install spacy && python -m spacy download en
+nlp = spacy.load('en')
+
+from sklearn.decomposition import TruncatedSVD
 
 '''Please feel free to modify this function (load_models)
    Make sure the function return the models required for the function below evaluate model
@@ -139,12 +149,31 @@ def evaluate_model(model, query, document):
     # result = model[1].predict([[cos_title, common_title, cos_body, common_body]])
     return result[0],0.5
 
-def common_terms(x,y):
+def common_terms(x,y, stopwords=False):
+    x = x.split(" ")
+    y = y.split(" ")
+    if stopwords:
+        x = [PorterStemmer().stem(t) for t in x]
+        y = [PorterStemmer().stem(t) for t in y]
+
     count = 0
-    for token in set(x.split(" ")):
-        if token in set(y.split(" ")):
+    for token in set(x):
+        if token in set(y) and token != "" and token != " ":
             count += 1
     return count
+def common_count(x,y, stopwords=False):
+    x = x.split(" ")
+    y = y.split(" ")
+    if stopwords:
+        x = [PorterStemmer().stem(t) for t in x]
+        y = [PorterStemmer().stem(t) for t in y]
+
+    count = 0
+    for token in x:
+        if token in y and token != "" and token != " ":
+            count += 1
+    return count
+
 
 def cosine_similarity(x,y, w2v=False):
     if x is None or y is None:
@@ -196,14 +225,33 @@ def convert_to_w2v(tokens, w2v):
 
 def convert_to_blob(text):
     opinion = TextBlob(text)
-    return opinion.sentiment
+    return opinion
 
 
-def naiveBayes(text):
-    opinion = TextBlob(text, analyzer=NaiveBayesAnalyzer())
-    print(opinion)
-    print(opinion.sentiment)
-    exit()
+def sentiment(textblob):
+    return textblob.sentiment
+
+
+def noun_counter(textblob):
+    return len(textblob.noun_phrases)
+def noun_counter_union(textblob_doc, textblob_query):
+    counter =0
+    noun1 = textblob_doc.noun_phrases
+    noun2 = textblob_query.noun_phrases
+    for n1 in noun1:
+        for n2 in noun2:
+            if n1 == n2:
+                counter +=1
+    return counter
+
+def noun_frequency(textblob):
+    return textblob.np_counts
+
+def polarity(textblob):
+    return textblob.polarity
+
+def subjectivity(textblob):
+    return textblob.subjectivity
 
 
 def cos_avg(x, y):
@@ -216,6 +264,13 @@ def split_tuples(tup):
         p.append(t[0])
         s.append(t[1])
     return p, s
+def spacy_similarity(doc, query):
+    doc1 = nlp(doc)
+    doc2 = nlp(query)
+    return doc1.similarity(doc2)
+
+
+
 
 
 def create_model(all_documents_file, relevance_file,query_file):
@@ -235,8 +290,8 @@ def create_model(all_documents_file, relevance_file,query_file):
 
     ''' Step 3. Creating a model for generating TF feature'''
 
+    # vectorizer = TfidfVectorizer()
     vectorizer = TfidfVectorizer( min_df=0.0, max_df=1.0, stop_words="english", lowercase=True, norm="l2", strip_accents='ascii')
-    # vectorizer = TfidfVectorizer(use_idf=True, sublinear_tf=True,smooth_idf=True, min_df=0.0, max_df=1.0, stop_words="english", lowercase=True, norm="l2", strip_accents='ascii')
     vectorizer = vectorizer.fit(rv["all_text"])
 
 
@@ -245,27 +300,29 @@ def create_model(all_documents_file, relevance_file,query_file):
     w2v_model.build_vocab(rv["all_text"])
 
 
-    # ''' Doc to Vec Model'''
-    # sentences = []
-    # for index, row in rv.iterrows():
-    #     sentences.append(LabeledSentence(row["all_text"].split(), ['SENT %s' % index]))
-    # d2v_model = gensim.models.Doc2Vec(sentences)
 
     ''' Step 4. Saving the model for TF features'''
     joblib.dump(vectorizer, 'resources/vectorizer.pkl')
     w2v_model.save('resources/w2v.model')
-    # d2v_model.save('resources/d2v.model')
 
     ''' Step 5. Converting query and title to vectors and finding cosine similarity of the vectors'''
     rv["query_vec"] = rv.apply(lambda x: vectorizer.transform([x["query"]]), axis =1)
     rv["doc_vec_title"] = rv.apply(lambda x: vectorizer.transform([x["title"] ]), axis =1)
     rv["doc_vec_body"] = rv.apply(lambda x: vectorizer.transform([ x["body"]]), axis =1)
 
+
+    rv["cosine_title"]  = rv.apply(lambda x: cosine_similarity(x['doc_vec_title'], x['query_vec']), axis=1)
+    rv["cosine_body"]  = rv.apply(lambda x: cosine_similarity(x['doc_vec_body'], x['query_vec']), axis=1)
+    rv["common_title"] = rv.apply(lambda x: common_terms(x["title"], x["query"]), axis =1)
+
+
     '''  COS and COMMON '''
     rv["cosine_title"]  = rv.apply(lambda x: cosine_similarity(x['doc_vec_title'], x['query_vec']), axis=1)
     rv["cosine_body"]  = rv.apply(lambda x: cosine_similarity(x['doc_vec_body'], x['query_vec']), axis=1)
     rv["common_title"] = rv.apply(lambda x: common_terms(x["title"], x["query"] ), axis =1)
-    rv["common_body"] = rv.apply(lambda x: common_terms(x["body"], x["query"] ), axis =1)
+    rv["common_body"] = rv.apply(lambda x: common_terms(x["body"], x["query"]), axis =1)
+    rv["common_title_s"] = rv.apply(lambda x: common_count(x["title"], x["query"],stopwords=True ), axis =1)
+    rv["common_body_s"] = rv.apply(lambda x: common_count(x["body"], x["query"], stopwords=True) , axis =1)
 
     '''  Word 2V'''
     rv["query_vec_w2v"] = rv.apply(lambda x: convert_to_w2v(x["query"], w2v_model), axis=1)
@@ -286,17 +343,36 @@ def create_model(all_documents_file, relevance_file,query_file):
     rv["doc_vec_blob"] = rv.apply(lambda x: convert_to_blob(x["title"]), axis=1)
     rv["body_vec_blob"] = rv.apply(lambda x: convert_to_blob(x["body"]), axis=1)
 
-    '''Split tuples'''
-    rv["query_vec_blob_p"],rv["query_vec_blob_s"]=  split_tuples(rv["query_vec_blob"])
-    rv["doc_vec_blob_p"],rv["doc_vec_blob_s"]=  split_tuples(rv["doc_vec_blob"])
-    rv["body_vec_blob_p"],rv["body_vec_blob_s"]=  split_tuples(rv["body_vec_blob"])
+    '''' Textblob noun count Union '''
+    rv["query_noun"] = rv.apply(lambda x: noun_counter( x["query_vec_blob"]), axis=1)
+    rv["title_noun"] = rv.apply(lambda x: noun_counter(x["doc_vec_blob"]), axis=1)
+    rv["body_noun"] = rv.apply(lambda x: noun_counter(x["body_vec_blob"]), axis=1)
 
-    '''Sentament cosine'''
-    rv["cosine_title_pblob"]  = rv.apply(lambda x: cosine_similarity(x['doc_vec_blob_p'], rv["query_vec_blob_p"], w2v=True), axis=1)
-    rv["cosine_title_sblob"]  = rv.apply(lambda x: cosine_similarity(x['doc_vec_blob_s'], rv["query_vec_blob_s"], w2v=True), axis=1)
-    rv["cosine_body_pblob"]  = rv.apply(lambda x: cosine_similarity(x['body_vec_blob_p'], x['query_vec_blob_p'], w2v=True), axis=1)
-    rv["cosine_body_sblob"]  = rv.apply(lambda x: cosine_similarity(x['body_vec_blob_s'], x['query_vec_blob_p'], w2v=True), axis=1)
 
+    '''' Textblob noun count Union '''
+    rv["title_noun_union"] = rv.apply(lambda x: noun_counter_union(x["doc_vec_blob"], x["query_vec_blob"]), axis=1)
+    rv["body_noun_union"] = rv.apply(lambda x: noun_counter_union(x["body_vec_blob"], x["query_vec_blob"]), axis=1)
+
+
+    '''' polarity '''
+    rv["query_polarity"] = rv.apply(lambda x: polarity(x["query_vec_blob"]), axis=1)
+    rv["title_polarity"] = rv.apply(lambda x: polarity(x["doc_vec_blob"]), axis=1)
+    rv["body_polarity"] = rv.apply(lambda x: polarity(x["body_vec_blob"]), axis=1)
+
+    '''' subjectivity '''
+    rv["query_subjectivity"] = rv.apply(lambda x: subjectivity(x["query_vec_blob"]), axis=1)
+    rv["title_subjectivity"] = rv.apply(lambda x: subjectivity(x["doc_vec_blob"]), axis=1)
+    rv["body_subjectivity"] = rv.apply(lambda x: subjectivity(x["body_vec_blob"]), axis=1)
+
+
+
+    '''cosine portability'''
+    rv["cosine_title_polarity"]  = rv.apply(lambda x: cosine_similarity(x['title_polarity'], rv["query_polarity"], w2v=True), axis=1)
+    rv["cosine_body_polarity"]  = rv.apply(lambda x: cosine_similarity(x['body_polarity'], x['query_polarity'], w2v=True), axis=1)
+
+    ''' cosine subjectivity'''
+    rv["cosine_title_subjectivity"]  = rv.apply(lambda x: cosine_similarity(x['title_subjectivity'], rv["query_subjectivity"], w2v=True), axis=1)
+    rv["cosine_body_subjectivity"]  = rv.apply(lambda x: cosine_similarity(x['body_subjectivity'], x['query_subjectivity'], w2v=True), axis=1)
 
 
     ''' IDF Extraction'''
@@ -326,30 +402,19 @@ def create_model(all_documents_file, relevance_file,query_file):
     rv["prob_body_idf"] = rv.apply(lambda x: idf_prob(x["doc_vec_body"]), axis =1)
     rv["mean_body_idf"] = rv.apply(lambda x: np.mean(x["doc_vec_body"]), axis =1)
 
+    ''' Cos Meta data'''
+    rv["cosine_title_max_idf"]  = rv.apply(lambda x: cosine_similarity(x['max_title_idf'], rv["max_query_idf"], w2v=True), axis=1)
+    rv["cosine_body_max_idf"]  = rv.apply(lambda x: cosine_similarity(x['max_body_idf'], rv["max_query_idf"], w2v=True), axis=1)
 
 
-    ''' W2V Data '''
-    rv["query_vec_w2v_pos"] = rv.apply(lambda x: np.argmax(x["query_vec"]), axis=1)
-    rv["query_vec_w2v_max"] = rv.apply(lambda x: np.max(x["query_vec"]), axis=1)
-    rv["query_vec_w2v_min"] = rv.apply(lambda x: np.min(x["query_vec"]), axis=1)
-    rv["query_vec_w2v_sum"] = rv.apply(lambda x: np.sum(x["query_vec"]), axis=1)
-    rv["query_vec_w2v_norm"] = np.divide(rv["query_vec_w2v_sum"] ,rv["len_title_idf"] )
+    rv["cosine_norm_title_idf"]  = rv.apply(lambda x: cosine_similarity(x['norm_title_idf'], rv["norm_query_idf"], w2v=True), axis=1)
+    rv["cosine_norm_body_idf"]  = rv.apply(lambda x: cosine_similarity(x['norm_body_idf'], rv["norm_body_idf"], w2v=True), axis=1)
 
+    rv["cosine_mean_title_idf"]  = rv.apply(lambda x: cosine_similarity(x['mean_title_idf'], rv["mean_query_idf"], w2v=True), axis=1)
+    rv["cosine_mean_body_idf"]  = rv.apply(lambda x: cosine_similarity(x['mean_body_idf'], rv["mean_query_idf"], w2v=True), axis=1)
 
-    rv["doc_vec_w2v_pos"] = rv.apply(lambda x: np.argmax(x["doc_vec_w2v"]), axis=1)
-    rv["doc_vec_w2v_max"] = rv.apply(lambda x: np.max(x["doc_vec_w2v"]), axis=1)
-    rv["doc_vec_w2v_min"] = rv.apply(lambda x: np.min(x["doc_vec_w2v"]), axis=1)
-    rv["doc_vec_w2v_sum"] = rv.apply(lambda x: np.sum(x["doc_vec_w2v"]), axis=1)
-    rv["doc_vec_w2v_norm"] = np.divide(rv["doc_vec_w2v_sum"] ,rv["len_title_idf"] )
-
-
-    rv["body_vec_w2v_pos"] = rv.apply(lambda x: np.argmax(x["body_vec_w2v"]), axis=1)
-    rv["body_vec_w2v_max"] = rv.apply(lambda x: np.max(x["body_vec_w2v"]), axis=1)
-    rv["body_vec_w2v_min"] = rv.apply(lambda x: np.min(x["body_vec_w2v"]), axis=1)
-    rv["body_vec_w2v_sum"] = rv.apply(lambda x: np.sum(x["body_vec_w2v"]), axis=1)
-    rv["body_vec_w2v_norm"] = np.divide(rv["doc_vec_w2v_sum"] ,rv["len_title_idf"] )
-
-
+    rv["cosine_prob_title_idf"]  = rv.apply(lambda x: cosine_similarity(x['mean_title_idf'], rv["mean_query_idf"], w2v=True), axis=1)
+    rv["cosine_prob_body_idf"]  = rv.apply(lambda x: cosine_similarity(x['mean_body_idf'], rv["mean_query_idf"], w2v=True), axis=1)
 
 
 
@@ -357,126 +422,52 @@ def create_model(all_documents_file, relevance_file,query_file):
 
 
     ''' Step 6. Defining the feature and label  for classification'''
-    X = rv[ ["common_title"] +  ["common_body"]
-        + ["max_query_idf"]  + ["max_pos_query_idf"]  + ["norm_query_idf"] + ["len_query_idf"] + ["prob_query_idf"]+["mean_query_idf"]
-        + ["max_title_idf"]  + ["max_pos_title_idf"]  + ["norm_title_idf"] + ["len_title_idf"] + ["prob_title_idf"]+["mean_title_idf"]
-        + ["max_body_idf"]   + ["max_pos_body_idf"]   + ["norm_body_idf"]  + ["len_body_idf"]  + ["prob_body_idf"] +["mean_body_idf"]
-        # + ["query_vec_w2v_pos"] + ["doc_vec_w2v_pos"] + ["body_vec_w2v_pos"]
-
-    ]
-
-
-    # X = rv[ ["avg_cos_title"] + ["cosine_title"]+ ["cosine_title_w2v"]+["common_title"] + ["avg_cos_body"]+ ["cosine_body"] + ["cosine_body_w2v"] + ["common_body"]
-    #     + ["cosine_title_pblob"] + ["cosine_title_sblob"] + ["cosine_body_pblob"] + ["cosine_body_sblob"]
-    #     + ["max_query_idf"]  + ["max_pos_query_idf"]  + ["norm_query_idf"] + ["len_query_idf"] + ["prob_query_idf"]
-    #     + ["max_title_idf"]  + ["max_pos_title_idf"]  + ["norm_title_idf"] + ["len_title_idf"] + ["prob_title_idf"]
-    #     + ["max_body_idf"]   + ["max_pos_body_idf"]   + ["norm_body_idf"]  + ["len_body_idf"]  + ["prob_body_idf"]
-    #     + ["query_vec_w2v_pos"] + ["query_vec_w2v_max"]   + ["query_vec_w2v_norm"] + ["query_vec_w2v_sum"]
-    #     + ["doc_vec_w2v_pos"] + ["doc_vec_w2v_max"] +   ["doc_vec_w2v_norm"] + ["doc_vec_w2v_sum"]
-    #     + ["body_vec_w2v_pos"] + ["body_vec_w2v_max"] + ["body_vec_w2v_norm"] + ["body_vec_w2v_sum"]
-    #     + ["query_vec_blob_p"] + ["query_vec_blob_s"]
-    #     + ['doc_vec_blob_p'] + ['doc_vec_blob_s']
-    #     + ['body_vec_blob_p'] + ['body_vec_blob_s']
-    #
-    #     ]
+    X = rv[["cosine_title"] + ["cosine_body"]
+           + ["cosine_title_w2v"]+ ["cosine_body_w2v"]
+           + ["cosine_title_polarity"] + ["cosine_title_subjectivity"] + ["cosine_body_polarity"] + ["cosine_body_subjectivity"]
+           + ["max_query_idf"] + ["max_pos_query_idf"] + ["norm_query_idf"] + ["len_query_idf"] + ["prob_query_idf"]+["mean_query_idf"]
+           + ["max_title_idf"] + ["max_pos_title_idf"] + ["norm_title_idf"] + ["len_title_idf"] + ["prob_title_idf"]+["mean_title_idf"]
+           + ["max_body_idf"] + ["max_pos_body_idf"] + ["norm_body_idf"] + ["len_body_idf"] + ["prob_body_idf"]+["mean_body_idf"]
+           + ["common_title"]+ ["common_body"]
+           + ["cosine_title_polarity"]+ ["cosine_body_polarity"]
+           + ["cosine_title_subjectivity"]+ ["cosine_body_subjectivity"]
+           ]
 
 
 
 
 
-    # X = rv[ ["avg_cos_title"] + ["cosine_title"]+ ["cosine_title_w2v"]+["common_title"] + ["avg_cos_body"]+ ["cosine_body"] + ["cosine_body_w2v"] + ["common_body"]
-    #     + ["cosine_title_pblob"] + ["cosine_title_sblob"] + ["cosine_body_pblob"] + ["cosine_body_sblob"]
-    #     + ["max_query_idf"]  + ["max_pos_query_idf"]  + ["norm_query_idf"] + ["len_query_idf"] + ["prob_query_idf"]
-    #     + ["max_title_idf"]  + ["max_pos_title_idf"]  + ["norm_title_idf"] + ["len_title_idf"] + ["prob_title_idf"]
-    #     + ["max_body_idf"]   + ["max_pos_body_idf"]   + ["norm_body_idf"]  + ["len_body_idf"]  + ["prob_body_idf"]
-    #     + ["query_vec_w2v_pos"] + ["query_vec_w2v_max"] + ["query_vec_w2v_min"]  + ["query_vec_w2v_norm"] + ["query_vec_w2v_sum"]
-    #     + ["doc_vec_w2v_pos"] + ["doc_vec_w2v_max"] + ["doc_vec_w2v_min"]   + ["doc_vec_w2v_norm"] + ["doc_vec_w2v_sum"]
-    #     + ["body_vec_w2v_pos"] + ["body_vec_w2v_max"] + ["body_vec_w2v_min"] + ["body_vec_w2v_norm"] + ["body_vec_w2v_sum"]
-    #     + ["query_vec_blob_p"] + ["query_vec_blob_s"]
-    #     + ['doc_vec_blob_p'] + ['doc_vec_blob_s']
-    #     + ['body_vec_blob_p'] + ['body_vec_blob_s']
-    #
-    #     ]
-
-    # X = rv[ ["avg_cos_title"] + ["cosine_title"]+ ["cosine_title_w2v"]+["common_title"] + ["avg_cos_body"]+ ["cosine_body"] + ["cosine_body_w2v"] + ["common_body"]
-    #     + ["cosine_title_pblob"] + ["cosine_title_sblob"] + ["cosine_body_pblob"] + ["cosine_body_sblob"]
-    #     + ["max_query_idf"]  + ["max_pos_query_idf"]  + ["norm_query_idf"] + ["len_query_idf"] + ["prob_query_idf"]
-    #     + ["max_title_idf"]  + ["max_pos_title_idf"]  + ["norm_title_idf"] + ["len_title_idf"] + ["prob_title_idf"]
-    #     + ["max_body_idf"]   + ["max_pos_body_idf"]   + ["norm_body_idf"]  + ["len_body_idf"]  + ["prob_body_idf"]
-    #     + ["query_vec_w2v_pos"] + ["query_vec_w2v_max"]  + ["query_vec_w2v_min"] + ["query_vec_w2v_sum"]
-    #     + ["doc_vec_w2v_pos"] + ["doc_vec_w2v_max"]  + ["doc_vec_w2v_min"] + ["doc_vec_w2v_sum"]
-    #     + ["body_vec_w2v_pos"] + ["body_vec_w2v_max"] + ["body_vec_w2v_min"] + ["body_vec_w2v_sum"]
-    #
-    #     ]
-
-
-
-
-
-
-
-
-
-
-    # X = rv[ ["avg_cos_title"] + ["cosine_title"]+ ["cosine_title_w2v"]+["common_title"] + ["avg_cos_body"]+ ["cosine_body"] + ["cosine_body_w2v"] + ["common_body"]
-    #     + ["cosine_title_pblob"] + ["cosine_title_sblob"] + ["cosine_body_pblob"] + ["cosine_body_sblob"]
-    #     ]
-
-
-    # rv["title_vec_w2v_max"] = rv.apply(lambda x: np.max(x["query_vec"]), axis=1)
-    # rv["title_vec_w2v_min"] = rv.apply(lambda x: np.min(x["query_vec"]), axis=1)
-    # rv["doc_vec_w2v_max"] = rv.apply(lambda x: np.max(x["doc_vec_w2v"]), axis=1)
-    # rv["doc_vec_w2v_min"] = rv.apply(lambda x: np.min(x["doc_vec_w2v"]), axis=1)
-    # rv["body_vec_w2v_max"] = rv.apply(lambda x: np.max(x["body_vec_w2v"]), axis=1)
-    # rv["body_vec_w2v_min"]
-
-    from sklearn.preprocessing import normalize
-    # X = normalize(X)
-
-    # X = rv[
-    #      ["max_query_idf"]  + ["max_pos_query_idf"]  + ["norm_query_idf"] + ["len_query_idf"] + ["prob_query_idf"]
-    #     + ["max_title_idf"]  + ["max_pos_title_idf"]  + ["norm_title_idf"] + ["len_title_idf"] + ["prob_title_idf"]
-    #     + ["max_body_idf"]   + ["max_pos_body_idf"]   + ["norm_body_idf"]  + ["len_body_idf"]  + ["prob_body_idf"]]
-
-
-    # X = rv[ ["common_title"] + ["common_body"]]
-    #
-    # X = rv[ ["cosine_title"]+ ["common_title"] + ["cosine_body"] + ["common_body"]]
-
-    # print(rv["doc_vec_body"].shape)
-    # print(rv["doc_vec_body"][0].shape)
-    # # print([x for x in rv["doc_vec_body"]])
-    # # X = rv["doc_vec_body"].tomatrix
-    # X = np.zeros((len(rv["doc_vec_body"]), rv["doc_vec_body"][0].shape[1]))
-    # for i in range(len(rv["doc_vec_body"])):
-    #     doc = scipy.sparse.coo_matrix(rv["doc_vec_body"][i])
-    #     for doc_word, doc_idf in zip(doc.col, doc.data):
-    #         X[i][doc_word] = doc_idf
-    #
-    #
-    # print(X)
-
-    # Y = one_hot([v for k, v in rv["position"].items()])
     Y = [v for k, v in rv["position"].items()]
+
 
 
     ''' Step 7. Splitting the data for validation'''
     X_train, X_test, y_train, y_test = train_test_split(    X, Y, test_size = 0.33, random_state = 42)
 
+
+    from sklearn.preprocessing import StandardScaler
+    # scaler = StandardScaler()
+    # scaler = MinMaxScaler()
+    # scaler.fit(X_train)
+    # X_train = scaler.transform(X_train)
+    # X_test = scaler.transform(X_test)
+
+
+    #
+    # clf = VotingClassifier(
+    #     estimators=[('ab', AdaBoostClassifier()), ('gb',  GradientBoostingClassifier(n_estimators=100)) ],
+    #     # estimators=[('ab', AdaBoostClassifier()), ('gb',  GradientBoostingClassifier(n_estimators=100)), ('rf', RandomForestClassifier(n_estimators=100)) ],
+    #     voting='soft'
+    #
+    # ).fit(X_train, y_train)
+
+
     ''' Step 8. Classification and validation'''
     target_names = ['1', '2', '3','4']
 
-    from sklearn.svm import LinearSVC
 
 
-    # clf = RandomForestClassifier().fit(X_train, y_train)
-    clf = LinearSVC(random_state=0).fit(X_train, y_train)
-
-
-
-
-
-
+    clf = GradientBoostingClassifier(n_estimators=1000).fit(X_train, y_train)
 
 
 
@@ -485,27 +476,6 @@ def create_model(all_documents_file, relevance_file,query_file):
     ''' Step 9. Saving the data '''
     joblib.dump(clf, 'resources/classifier.pkl')
 
-#  Close but not using
-    # Y = [v for k, v in rv["position"].items()]
-    # Y = rv["position"]
-    # clf = LogisticRegression().fit(X_train, y_train)
-    # LogisticRegression(C=1.0, class_weight=None, dual=False, fit_intercept=True,
-    #                    intercept_scaling=1, penalty='l2', random_state=None, tol=0.0001)
-    # clf = RandomForestClassifier(class_weight="balanced").fit(X_train, y_train)
-    from sklearn.svm import LinearSVC
-
-    # clf = RandomForestClassifier(class_weight= [{0: 1, 1: 50}, {0: 1, 1: 25}, {0: 1, 1: 10}, {0: 1, 1: 5}] ).fit(X_train, y_train)
-
-    # clf = LinearSVC(random_state=0).fit(X_train, y_train)
-
-# NOT USING
-    # clf = MultinomialNB().fit(X_train, y_train)
-    # clf = OneVsRestClassifier(MultinomialNB()).fit(X_train, y_train)
-    # clf = MultinomialNB().fit(X_train, y_train)
-    # clf = RandomForestClassifier().fit(X_train, y_train)
-    from sklearn.svm import LinearSVC
-    # clf = OneVsRestClassifier(LinearSVC(random_state=0)).fit(X_train, y_train)
-    # clf = SVC().fit(X_train, y_train)
 
 
 if __name__ == '__main__':
